requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
langchain-groq>=0.1.5
groq>=0.5.0
tavily-python>=0.3.3
python-dotenv>=1.0.1
pydantic>=2.7.0
typing-extensions>=4.9.0